int => uLong

+++ dataplane/dataplaneconnection.cpp
@@ -15,8 +15,8 @@ void DataPlaneConnection::removeConnection() {
     BonjourSQL* qSql = BonjourSQL::getInstance();
 
     qDebug() << "Removing connection!";
-    qDebug() << friendUid<< "<" << qSql->getLocalUid();
-    if (friendUid.toInt() < qSql->getLocalUid().toInt()) { // friend is smaller, I am server
+    qDebug() << friendUid.toULong() << "<" << qSql->getLocalUid().toULong();
+    if (friendUid.toULong() < qSql->getLocalUid().toULong()) { // friend is smaller, I am server
         qDebug() << "friend is smaller!";
         if (client)
             client->stop();
